ractor pcapng into single module
